+++ paper/paper.bib
@@ -6,7 +6,8 @@
   volume={22},
   number={1},
   pages={84--97},
-  year={2005}
+  year={2005}, 
+  doi={10.1175/JTECH-1687.1}
 }
 
 @article{loder:1980,
@@ -16,7 +17,8 @@
   volume={10},
   number={9},
   pages={1399--1416},
-  year={1980}
+  year={1980}, 
+  doi={10.1175/1520-0485(1980)010<1399:TROTCO>2.0.CO;2}
 }
 
 @article{munk:1998,
@@ -27,7 +29,8 @@
   number={12},
   pages={1977--2010},
   year={1998},
-  publisher={Elsevier}
+  publisher={Elsevier}, 
+  doi={10.1016/S0967-0637(98)00070-3}
 }
 
 @article{padman:2004,
@@ -37,7 +40,8 @@
   volume={31},
   number={2},
   year={2004},
-  publisher={Wiley Online Library}
+  publisher={Wiley Online Library}, 
+  doi={10.1029/2003GL019003}
 }
 
 
@@ -49,7 +53,8 @@
   number={1},
   pages={142--184},
   year={2018},
-  publisher={Wiley Online Library}
+  publisher={Wiley Online Library}, 
+  doi={10.1002/2016RG000546}
 }
 
 @misc{padman:2022, 
@@ -68,7 +73,8 @@
   number={5},
   pages={399--413},
   year={2000},
-  publisher={Springer}
+  publisher={Springer}, 
+  doi={10.1007/s001900000101}
 }
 
 @article{stammer:2014,
@@ -79,5 +85,6 @@
   number={3},
   pages={243--282},
   year={2014},
-  publisher={Wiley Online Library}
+  publisher={Wiley Online Library}, 
+  doi={10.1002/2014RG000450}
 } 
