### 2023-10-05 #5770 @kthyng
Hi @chadagreene and sorry for the wait.

I want to state that @chadagreene and I are co-authors on a [paper from 2016](https://tos.org/oceanography/assets/docs/29-3_thyng.pdf) and [paper from 2019](https://agupubs.onlinelibrary.wiley.com/doi/pdfdirect/10.1029/2019GC008392). The work done for this latter paper was done in [January 2019](https://github.com/matplotlib/cmocean/commit/b48d8314252a4ef14bf8ef70e2b499ebef950df0) so I am ok to edit this submission in terms of a potential conflict of interest, as per our [policy](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html#joss-conflict-of-interest-policy).
