### 2023-10-13 #5770 @mhalvers
@kthyng I'm flattered that you thought of me!  I haven't coded in Matlab or worked in oceanography since 2020 🙁.  Out of curiosity I attempted to open Matlab on my computer and it crashed immediately.  I'm not sure how to fix the issue, and I also don't have access to newer versions.  In light of all I have to decline.
