### 2023-10-26 #5770 @kthyng
@jklymak do you still use Matlab? If so would you be interested in reviewing this submission to the Journal for Open Source Software?  Our reviews are done openly in a github issue and are checklist driven — you can get more information here: https://joss.readthedocs.io/en/latest/reviewer_guidelines.html#
I would ask that you start your review in the next 3 weeks with the intention of having it finished within 4-6 weeks, given that the process is often iterative with the submitting author(s).

Thanks for your consideration!
