### 2023-10-26 #5770 @kthyng
@mhalvers Yeah that is a good idea but I couldn't find a github account for him which would make jumping into a JOSS review pretty difficult.
