### 2023-11-06 #5770 @kthyng
@dgwyther Thank you! Yeah there isn't a good way in github to say you're out of office...

If you're interested you can register as a reviewer for JOSS here so that you're discoverable for next time there is a relevant submission: https://reviewers.joss.theoj.org/
