### 2023-11-20 #6018 @adamdanz
Thanks for the bump, Kristen.  I’ll get started on this this week and will let you know if I have any questions.

Adam

From: Kristen Thyng ***@***.***>
Sent: Monday, November 20, 2023 12:44 PM
Cc: Adam Danz ***@***.***>; Mention ***@***.***>
Subject: Re: [openjournals/joss-reviews] [REVIEW]: Tide Model Driver for MATLAB (Issue #6018)

@adamdanz<https://github.com/adamdanz> and @lnferris<https://github.com/lnferris> Any questions for getting your review started?
