### 2023-12-10 #6018 @adamdanz
I've completed this review.  The paper summarizes the software well; the documentation is superb and the tests are sufficient to confirm consistent and expected results.

I wasn't able to find explicit guidelines for 3rd party contributions and issue reports within this submission but I checked this box as complete since the submission is in a public GitHub repo that is commonly used for these 3rd party interactions.
