### 2023-12-17 chadagreene/Tide-Model-Driver#6 @lnferris
Community guidelines
This is part of [JOSS review](https://github.com/openjournals/joss-reviews/issues/6018).

Please add guidance (even a quick sentence) in README.md providing instructions for bug fixes or contributions. Will you be using GitHub issues to track errors? What kind of contributions are you accepting, if any? Is Chad, Svetlana, or someone else the primary point of contact?
