### 2023-12-17 chadagreene/Tide-Model-Driver#7 @lnferris
Summary
This is part of [JOSS review](https://github.com/openjournals/joss-reviews/issues/6018).

Please replace "meaningful form" (on line 20 of the paper) with specifically what type of parameter(s) the software produces in order to give the reader a preview of whether this toolbox is suitable for their purposes.
