### 2023-12-17 chadagreene/Tide-Model-Driver#8 @lnferris
Quality of writing
This is part of [JOSS review](https://github.com/openjournals/joss-reviews/issues/6018).

This paper (not the software) is low on content relative to other ocean-related JOSS papers and requires prior experience with TMD 2.5 in order to be at all meaningful. As the markdown is full of great documentation, I recommend spending a small amount of time repurposing an example and/or a few key function highlights and including it in your paper.
