### 2023-12-19 chadagreene/Tide-Model-Driver#9 @lnferris
Suggestion: tmd_conlist
This is part of [JOSS review](https://github.com/openjournals/joss-reviews/issues/6018).

If you end the function with something like:

```
A = strings(length(conList),1);
for i = 1:length(conList)
    A(i) = conList{i};
end
conList = A;
```

It will give the user with a string array, which might be preferable to a cell array.
