### 2023-12-19 chadagreene/Tide-Model-Driver#10 @lnferris
State of the field: other commonly-used packages
This is part of [JOSS review](https://github.com/openjournals/joss-reviews/issues/6018).

TMD 2.5 cites two alternative versions of the toolbox, Fortran (https://www.tpxo.net/otps) and Python (https://github.com/tsutterley/pyTMD). It is appropriate to mention these around line 27 of the paper, as you have done in README.md.

Similarly, it is probably appropriate to explain or change "The MATLAB version...no major updates...since its inception". The citation for TMD 2.5 is from 2022 which is inconsistent with the narrative that no one has touched it since its inception.
