### 2023-12-19 chadagreene/Tide-Model-Driver#11 @lnferris
Functionality documentation: functions renamed without documentation
This is part of [JOSS review](https://github.com/openjournals/joss-reviews/issues/6018).

In README.md, "Most functions were renamed and inputs have changed to make them more intuitive and easier to use."

Many users of TMD 2.5 do not use the functions as standalone functions; rather, they are called by other tools to perform tasks such as preparing tidal inputs for circulation models. Before incorporating TMD 3.0 into their processing pipeline, the large user base of TMD (mentioned in your paper) will have to spend significant time searching through the documentation and source code to determine: what actually changed (incl. any bug fixes), what was just renamed, and what has subtly different outputs (such as cm/s vs. m/s, terminology "Depth" vs. "Elevation" vs. "Z" vs. "zi" vs. "z" vs. "TS" vs. 'h', "amp" vs. '*Am'). For example, I can get same thing using:

```
% TMD 2.5
filename = '/Users/lnferris/MD_Matlab_Toolbox_v2.5-master/TMD/DATA/Model_tpxo7.2';
[TS,~]=tmd_tide_pred(filename,time,lat,lon,'z',[]);

% TMD 3.0
filename = '/Users/lnferris/TPXO9_atlas_v5.nc';
z = tmd_predict(filename,lat,lon,time,'h');

figure
plot(time,z)
hold on
plot(time,TS)
```
I want the authors to create a chart (or other medium of their choice) that outlines which TMD 2.5 functions have direct proxies to TMD 3.0, which functions were slightly-but-critically changed, and which functions are completely novel. The purpose of this is to minimize confusion as people adopt the new version.
