### 2023-12-20 chadagreene/Tide-Model-Driver#12 @lnferris
Suggestion: TPXO coming soon
This is part of [JOSS review](https://github.com/openjournals/joss-reviews/issues/6018).

TPXO is a staple, and so I tried running TPXO9_atlas_v5_to_NetCDF.m and it produced the NetCDF file, but it took hours to run. Since you have the code already, would it be possible to supply the file (via Box, etc.) to users? Is the file "coming soon" an issue with QAQC or a permissions? If it is a permissions issue, could you mention the existence of TPXO9_atlas_v5_to_NetCDF.m next to "data access" on tide_model_data.md, in the interim?
