### 2023-12-20 chadagreene/Tide-Model-Driver#13 @lnferris
tmd_predict: documenting input time and lon
This is part of [JOSS review](https://github.com/openjournals/joss-reviews/issues/6018).

Be explicit that Matlab datenum means days since 00-Jan-0000 because an inexperienced user might misinterpret the provided example as requesting days since 1800-01-01.

Mention that -180/180 or 0/360 format is accepted.
