### 2023-12-20 chadagreene/Tide-Model-Driver#14 @lnferris
Suggestion: tide_model_data.md
This is part of [JOSS review](https://github.com/openjournals/joss-reviews/issues/6018).

Change "complex heights and transport coefficients" to "complex heights (h) and transport coefficients (for u, v, U, V)" where appropriate.

Or specify in the tmd_predict documentation that having the transport coefficients (or not) changes the allowed value of `ptype` ; and make a note when describing inputs in the tmd_ellipse documentation that this is for models with transport coefficients only.
