### 2023-12-20 chadagreene/Tide-Model-Driver#15 @lnferris
Automated Tests: tide_model_intercomparison.md
This is part of [JOSS review](https://github.com/openjournals/joss-reviews/issues/6018) and related to [Issue 11](https://github.com/chadagreene/Tide-Model-Driver/issues/11)

It would be nice to also include a test(s) comparing timeseries produced using the TMD 2.5 vs. TMD 3.0 functions. I don't think these tests needs to be automated; a code block would be fine.
