### 2023-12-20 chadagreene/Tide-Model-Driver#16 @lnferris
tmd_data_documentation.md
This is part of [JOSS review](https://github.com/openjournals/joss-reviews/issues/6018).

Define x and y
