### 2023-12-21 chadagreene/Tide-Model-Driver#17 @lnferris
unexpected behavior: tmd_interp and tmd_data
This is part of [JOSS review](https://github.com/openjournals/joss-reviews/issues/6018).

There are two issues here: the toolbox seems to have inconsistent longitudinal range handling (whether -180/180 or 0/360 is fine), and the inability to interpolate data in a region which should be available.  (If I am misunderstanding the usage of tmd_interp and this is producing the latter problem, please let me know.)

Function tmd_interp seems to work for TPXO but not Arc2km

```
filename = '/Users/lnferris/Desktop/TPXO9_atlas_v5.nc';
%filename = '/Users/lnferris/Desktop/Arc2kmTM_v1.nc';

latq = linspace(70,72,60*2);
lonq = linspace(-66,-62,60*4);
[loni,lati] = meshgrid(lonq,latq);
zi = tmd_interp(filename,variable,loni,lati);
```
and produces the following error for Arc2km:

```
Error using tmd_data
No tide data available in the region of interest.

Error in tmd_interp (line 160)
      [z,x,y] = tmd_data(filename,variable,'bounds',[xi(:) yi(:)],varargin{:});

Error in check_TBD (line 83)
zi = tmd_interp(filename,variable,loni,lati);

```
I tried to use tmd_data to confirm the data exists...

```
conid = 3;
variable = 'hRe';
[Z,lon,lat] = tmd_data(filename,variable,'geo');
ocean = tmd_data(filename,'mask');
data = Z(:,:,conid);
data(~ocean) = NaN;

figure
pcolor(lon,lat,data);
shading flat
hold on
xline(-66); xline(-62); yline(70); yline(72)
xlim([-80 -50]); ylim([60 80])
```
and got this strange result for Arc2km (e.g. what are the stripes/continuities?):

and this for TPXO (no negative longitude handling):
