### 2023-12-21 chadagreene/Tide-Model-Driver#18 @lnferris
References: in code but not paper
This is part of [JOSS review](https://github.com/openjournals/joss-reviews/issues/6018).

Snyder, J. P., 1982
Foreman, M. G. G., & Henry, R. F. (1989)
