### 2023-12-21 chadagreene/Tide-Model-Driver#19 @lnferris
Minor documentation questions
This is part of [JOSS review](https://github.com/openjournals/joss-reviews/issues/6018).

[pu,pf] = tmd_nodal(constituents,p,N): what are pu, pf?

tmd_ll2ps(lat,lon,SLAT,SLON,HEMI): allowed inputs (vector, array, etc.)?

tmd_harp(t,hc,constituents,p,N,ph,omega): physical units of input/outputs?

dh = tmd_InferMinor(hc,constituents,t,s,h,p,N): what is a "correction"?

tmd_astrol.m: why are there question marks on line 51?  Given there are several time formats within TMD stack code: days since Jan 1, 1992; days since Nov 17, 1858, and MATLAB datenum (days since 00-Jan-0000'); the "???" does not inspire user confidence.
