### 2023-12-21 chadagreene/Tide-Model-Driver#6 @chadagreene
Good idea. I've added a section in the README on how to contribute. As for the personal point of contact, I've been careful not to encourage too many personal private communications bc in my experience that really encourages emails from strangers asking me to do their homework or research for them :-)
