### 2023-12-21 chadagreene/Tide-Model-Driver#20 @lnferris
TMD 2.5 to TMD 3.0 question
This is part of [JOSS review](https://github.com/openjournals/joss-reviews/issues/6018) and related to [Issue 11](https://github.com/chadagreene/Tide-Model-Driver/issues/11).

This is testing a function call (`tmd_extract_HC()`) used in a common tidal preparation for ROMS ([otps2frc_v5.m](https://www.myroms.org/hunter/roms_tides/update/otps/)). I get that `u_phase `and `uPh` are not pseudo-equivalent. Is there a units mismatch going on?

```
filename='/Users/lnferris/TMD_Matlab_Toolbox_v2.5-master/TMD/DATA/Model_tpxo7.2';
[z_amp,z_phase,Depth,conList]=tmd_extract_HC(filename,lat,lon,'z',[]);
[u_amp,u_phase,~,~]=tmd_extract_HC(filename,lat,lon,'u',[]);

con = 1; % m2 from conList
figure
subplot(3,2,1)
pcolor(lon,lat,squeeze(z_amp(con,:,:)));
shading flat
colorbar
subplot(3,2,2)
pcolor(lon,lat,squeeze(z_phase(con,:,:)));
shading flat
colorbar
subplot(3,2,3)
pcolor(lon,lat,Depth);
shading flat
colorbar
subplot(3,2,4)
pcolor(lon,lat,squeeze(u_amp(con,:,:))./100);  %cm/s to m/s
shading flat
colorbar
subplot(3,2,5)
pcolor(lon,lat,squeeze(u_phase(con,:,:)));
shading flat
colorbar

%% TMD 3.0

addpath(genpath('/Users/lnferris/Desktop/Tide-Model-Driver-main'));
filename = '/Users/lnferris/Desktop/TPXO9_atlas_v5.nc';
conList = tmd_conlist(filename);
hAm = tmd_interp(filename,'hAm',lat,lon);
hPh = tmd_interp(filename,'hPh',lat,lon);
uAm = tmd_interp(filename,'uAm',lat,lon);
uPh = tmd_interp(filename,'uPh',lat,lon);
wct = tmd_interp(filename,'wct',lat,lon);

con = 4; % m2 from conList
figure
subplot(3,2,1)
pcolor(lon,lat,squeeze(hAm(:,:,con)));
shading flat
colorbar
subplot(3,2,2)
pcolor(lon,lat,squeeze(hPh(:,:,con)));
shading flat
colorbar
subplot(3,2,3)
pcolor(lon,lat,wct);
shading flat
colorbar
subplot(3,2,4)
pcolor(lon,lat,squeeze(uAm(:,:,con)));
shading flat
colorbar
subplot(3,2,5)
pcolor(lon,lat,squeeze(uPh(:,:,con)));
shading flat
colorbar
```
