### 2023-12-21 #6018 @lnferris
I am finished with my review. Overall it is a well-prepared software package and its existence will make tidal analysis more accessible to several fields (geology, oceanography, climate). Even though it is a revision of an older package (TMD 2.5), I believe the old package was never formally published in a journal; and this more streamlined revision (TMD 3.0) includes all of the original TMD 2.5 authors - so I think that is fine.  The linked issues are important to address, but I would be happy to see it published if they are successfully addressed.
