+++ paper/paper.bib
@@ -10,6 +10,18 @@
   doi={10.1175/JTECH-1687.1}
 }
 
+@article{foreman:1989,
+  title={The harmonic analysis of tidal model time series},
+  author={Foreman, Michael George G and Henry, Richard F},
+  journal={Advances in water resources},
+  volume={12},
+  number={3},
+  pages={109--120},
+  year={1989},
+  publisher={Elsevier}, 
+  doi={10.1016/0309-1708(89)90017-1}
+}
+
 @article{loder:1980,
   title={{Topographic rectification of tidal currents on the sides of Georges Bank}},
   author={Loder, John W},
