### 2023-12-22 chadagreene/Tide-Model-Driver#8 @chadagreene
We understand the need to ensure that the paper communicates what TMD3.0 does, without expecting readers to have prior experience with previous versions of TMD. We believe we've explained it well in the paper:

> The Tide Model Driver for MATLAB version 3.0 (TMD3.0) allows users to access and calculate tidal predictions from model coefficients at arbitrary locations and times, and this version of the software represents decades of development by the tide community...TMD3.0 is issued as a free, open-source software package that is available to all.

Regarding the suggestion to expand the paper to include code examples in the paper, the [JOSS guidelines](https://joss.readthedocs.io/en/latest/review_criteria.html) state:

> ...the JOSS paper (the compiled PDF associated with this submission) should only include:
>
> * A list of the authors of the software and their affiliations.
> * A summary describing the high-level functionality and purpose of the software for a diverse, non-specialist audience.
> * A clear statement of need that illustrates the purpose of the software.
> * A description of how this software compares to other commonly-used packages in this research area.
> * Mentions (if applicable) of any ongoing research projects using the software or recent scholarly publications enabled by it.
> * A list of key references including a link to the software archive.

The guidelines also state:

> Note the paper should not include software documentation...

And that makes sense, because syntax or best practices may evolve in the future, and the living documentation can always reflect any changes, but an example baked into the published paper would not be able to be updated.

Also, a little anecdote to describe a major reason we chose JOSS: In previous experiences publishing on software in traditional examples, we've been compelled to include examples, and that always presented a quandary. Any single example would be somewhat arbitrary and couldn't fully represent the scope of a toolbox. The temptation is to craft an elaborate example that showcases all the major functionality of the toolbox and shows how functions work together, but that would push readers directly into the deep end and would likely drown them in the unfamiliar syntax and function calls. Taking a more humane approach, we could craft a beginner's example of a simple tidal prediction, but even a simple example quickly becomes a tutorial if presented properly. For these reasons, examples in a static document are a bit doomed, which is likely why the JOSS style guide does  not allow for them.
