### 2023-12-22 chadagreene/Tide-Model-Driver#9 @chadagreene
The suggested approach is reasonable, but we don't see any clear benefit of string arrays over cell arrays, and they're easy enough to convert back and forth, so we leave this one as is.
