### 2023-12-22 chadagreene/Tide-Model-Driver#18 @chadagreene
We've added a citation to Foreman 1989 to the paper. Snyder is a reference to a specific polar stereographic map projection, and doesn't really belong in the main paper where the JOSS guidelines state that the content should provide only a high-level overview.
