### 2023-12-22 chadagreene/Tide-Model-Driver#12 @chadagreene
Thanks for this comment. The issue of data ownership is a delicate one, and we're working toward a solution that will keep everyone happy. The TPXO9_atlas_v5_to_NetCDF.m file is available for users who are particularly dedicated, but we understand the burden it places on users. That script was originally intended to be run just once and only by me, because it's slow and requires downloading and using the rats' nest of code that is TMD2.5. We don't really want to encourage users to use TPXO9_atlas_v5_to_NetCDF.m because it requires use of a toolbox we don't own or have control over, and the reason for creating TMD3.0 was to eliminate this exact type of confusing and unnecessary effort. I could go on about this, but suffice it to say, we understand your point well and we're working hard to find a way for users to access the data without so much friction. Until then, we just have to say it's Coming Soon.
