### 2023-12-22 8143a20
Issue #16
Adding a description of x,y in response to Issue #16.
