### 2023-12-22 9700d29
Update tide_model_data.md
Documentation update to address Issue #14.
