+++ doc/tide_model_data.md
@@ -8,7 +8,7 @@ Barotropic ocean models contain complex tidal coefficients for water height and 
 
 ### Global Ocean Models
 * **`TPXO9_atlas_v5.nc`** by Gary Egbert and Svetlana Erofeeva.
-	* **Variables:** Complex height & transport coefficients.
+	* **Variables:** Complex heights (h) and transport coefficients (for u, v, U, V).
 	* **Resolution:** 1/30 degree.
 	* **Constituents:** 15: 2n2 k1 k2 m2 m4 mf mm mn4 ms4 n2 o1 p1 q1 s1 s2.
 	* **Restrictions:** For noncommercial use only (contact Gary Egbert or Svetlana Erofeeva otherwise). 
@@ -17,7 +17,7 @@ Barotropic ocean models contain complex tidal coefficients for water height and 
 	* **Data Citation:** Egbert, G. D., and Erofeeva, S. Y. (2002). Efficient inverse modeling of barotropic ocean tides. *Journal of Atmospheric and Oceanic Technology* 19(2), 183-204.
 	
 * **`EOT20_ocean.nc`** by Hart-Davis et al. EOT20 is the latest in a series of global ocean tide models developed by *Hart-Davis et al*. at DGFI-TUM. EOT20 is created using residual tidal analysis of multi-mission altimetry data from 1992-2019. Eleven satellite altimetry missions are used and the FES2014b tide model used as the reference model for the residual tidal analysis. The model extends throughout the global ocean, with EOT20 ranging from 66°S to 66°N with FES2014b being used to fill in the higher latitudes.
-	* **Variables:** Complex height only.
+	* **Variables:** Complex height (h) only.
 	* **Resolution:** 1/8 degree.
 	* **Constituents:** 17: 2n2 j1 k1 k2 m2 m4 mf mm n2 o1 p1 q1 s1 s2 sa ssa t2.
 	* **Restrictions:** Users must cite Hart-Davis et al., 2021. 
@@ -28,7 +28,7 @@ Barotropic ocean models contain complex tidal coefficients for water height and 
 
 ### Regional Ocean Models 
 * **`CATS2008_v2023.nc`** Circum-Antarctic Tidal Simulation by Lana Erofeeva, Chad Greene, Susan Howard, and Laurie Padman. Updates from the original CATS2008 are described [here](cats2008_updates.md). 
-	* **Variables:** Complex height & transport coefficients.
+	* **Variables:** Complex heights (h) and transport coefficients (for u, v, U, V).
 	* **Resolution:** 2 km.
 	* **Constituents:** 10: m2 s2 n2 k2 k1 o1 p1 q1 mf mm.
 	* **Restrictions:** This work is licensed under the Creative Commons Attribution 4.0 International License. 
@@ -37,7 +37,7 @@ Barotropic ocean models contain complex tidal coefficients for water height and 
 	* **Data Citation:** .
 	
 * **`Gr1kmTM_v1.nc`** by Susan Howard and Laurie Padman is a regional model of Greenland, developed using the Regional Ocean Modeling System (ROMS) on a standard north polar stereographic projection. 
-	* **Variables:** Complex height & transport coefficients.
+	* **Variables:** Complex heights (h) and transport coefficients (for u, v, U, V).
 	* **Resolution:** 1 km.
 	* **Constituents:** 8: m2 s2 k1 o1 n2 p1 k2 q1.
 	* **Restrictions:** This work is licensed under the Creative Commons Attribution 4.0 International License.. 
@@ -46,7 +46,7 @@ Barotropic ocean models contain complex tidal coefficients for water height and 
[…]
