### 2023-12-22 6510e8a
Issue #13
Documentation edits to address Issue #13.
