### 2023-12-22 chadagreene/Tide-Model-Driver#19 @chadagreene
We've made changes to the function headers to clarify some of the points above, but we should note: These functions are not user-facing, and should not be interacted with directly. Accordingly, the bar for documentation needs only to meet that of `stream3c` or any of the other hundreds of official MATLAB functions you'll find under the hood.

And we agree, the mish-mash of time formats in some of the underlying functions is lousy coding practice and it's just begging to introduce errors (or at least erode confidence in the code, as you point out). However, these are the exact equations that have been employed in all previous versions of TMD and we felt it was best not to mess with them.
