### 2023-12-23 81de4b8
joss review
Updates in reference to Issue #11 and Issue #15.
