### 2023-12-23 chadagreene/Tide-Model-Driver#15 @chadagreene
Nice idea. We've now added a page that describes updates in TMD3.0 compared to previous versions, in which we point users to a script they may use to compare results: https://github.com/chadagreene/Tide-Model-Driver/blob/main/doc/whats_new.md
