### 2023-12-23 chadagreene/Tide-Model-Driver#11 @chadagreene
I'm embarrassed we didn't think of this before. We've now added a page that describes what's new in TMD3.0, including the suggested table: https://github.com/chadagreene/Tide-Model-Driver/blob/main/doc/whats_new.md#converting-from-previous-tmd-releases
