### 2023-12-23 chadagreene/Tide-Model-Driver#20 @chadagreene
@lnferris I don't have quite enough to replicate your code, and I'm not sure what difference you're seeing between the two results. Can you please upload the resulting figures or describe the differences you're seeing?
