### 2023-12-27 chadagreene/Tide-Model-Driver#17 @chadagreene
Whoa, @lnferris thanks for conducting such a close examination of the function behavior using multiple different models. This kind of test really helps us identify unexpected behavior and understand how users are using the functions.

You've identified two issues here:

### 1. Function `tmd_interp` seems to work for TPXO but not Arc2km

Thanks for providing a minimal working example--It just looks like you've swapped the order of lat and lon in `tmd_interp`. Here's a comparison of the two models using your lat,lon grid:
```matlab
latq = linspace(70,72,60*2);
lonq = linspace(-66,-62,60*4);
[loni,lati] = meshgrid(lonq,latq);

zi_tpx = tmd_interp('TPXO9_atlas_v5.nc','hAm',lati,loni,'con','m2');
zi_arc = tmd_interp('Arc2kmTM_v1.nc','hAm',lati,loni,'con','m2');

figure
subplot(1,2,1)
pcolor(loni,lati,zi_tpx);
shading flat
title('TPXO9_atlas_v5.nc m2 hAm','interpreter','none')
caxis([0 0.27])

subplot(1,2,2)
pcolor(loni,lati,zi_arc);
shading flat
title('Arc2kmTM_v1.nc m2 hAm','interpreter','none')
caxis([0 0.27])
```
![tide_test](https://github.com/chadagreene/Tide-Model-Driver/assets/15901019/c19ebca8-2eae-4ee3-b265-23f6a556108e)

There's actually an `islatlon` test on line 72 of `tmd_interp`, which usually catches potential mixups in the order, but coincidentally in this particular example, all of the lat and lon values are within the possible range of values that could be associated with latitudes or longitudes.

### 2. Stripey gridded data in Arc2km

I am able to replicate the error, and this is only an issue of how `pcolor` tries to display a warped grid. Here's a comparison of displaying the same data with `pcolor` versus `scatter` (although scatter is slow, so watch out).

```matlab
filename = 'Arc2kmTM_v1.nc';

conid = 3;
variable = 'hRe';
[Z,lon,lat] = tmd_data(filename,variable,'geo');
ocean = tmd_data(filename,'mask');
data = Z(:,:,conid);
data(~ocean) = NaN;

figure
subplot(1,2,1)
pcolor(lon,lat,data);
shading flat
hold on
xline(-66); xline(-62); yline(70); yline(72)
xlim([-80 -50]); ylim([60 80])
title pcolor

subplot(1,2,2)
scatter(lon(:),lat(:),5,data(:),'filled');
shading flat
hold on
xline(-66); xline(-62); yline(70); yline(72)
xlim([-80 -50]); ylim([60 80])
title scatter
```
![tide_test3](https://github.com/chadagreene/Tide-Model-Driver/assets/15901019/5b353d02-4ffd-408f-afda-e91cbcc4d28d)
