### 2024-01-02 chadagreene/Tide-Model-Driver#20 @kthyng
@lnferris Can you respond here?
