### 2024-01-03 chadagreene/Tide-Model-Driver#12 @lnferris
I understand that you cannot provide the .nc file for now, but would like you to mention the option of using TPXO9_atlas_v5_to_NetCDF.m (if it indeed produces a correct .nc file) next to Coming Soon, because an update to TMD which doesn't support the most important tidal model (TPXO) is probably less useful than TMD2.5.
