### 2024-01-03 chadagreene/Tide-Model-Driver#11 @lnferris
I think the proxy of tmd_interp is tmd_extract_HC
