### 2024-01-03 chadagreene/Tide-Model-Driver#17 @lnferris
My understanding is that this issue arises when data is not sorted with respect to x and y. Can you fix it by adding a line of code that sorts it?
