### 2024-01-03 chadagreene/Tide-Model-Driver#20 @lnferris
[lat.csv](https://github.com/chadagreene/Tide-Model-Driver/files/13823087/lat.csv)
[lon.csv](https://github.com/chadagreene/Tide-Model-Driver/files/13823088/lon.csv)
![2 5](https://github.com/chadagreene/Tide-Model-Driver/assets/24570061/ad5c3565-da0a-4ab0-af11-24e039041178)
![3 0](https://github.com/chadagreene/Tide-Model-Driver/assets/24570061/e87b7d96-d3a6-4f52-944d-8cb531c2c06c)
