### 2024-01-03 #6018 @lnferris
I 'checked' some items. Issues 12, 19, 20 are open in my opinion. I am not going to weigh in on whether issue 8 should still be open - the editor knows better than I whether this paper is adequate to illustrate the purpose of the software by JOSS standards.
