### 2024-01-04 15045f0
Update whats_new.md
Addressing Issue #11.
