### 2024-01-04 5c098bd
Update tide_model_data.md
Addressing Issue #12.
