### 2024-01-04 chadagreene/Tide-Model-Driver#12 @chadagreene
Good idea. I've made the note as suggested.
