### 2024-01-04 chadagreene/Tide-Model-Driver#17 @chadagreene
@lnferris I'm afraid this is just the nature of map projections. The raw data comes on a regular grid in projected space. The `'geo'` option provides lat,lon coordinates corresponding to each pixel in the raw data grid, so by nature the outputs can't really be sorted. If a regular lat,lon grid is desired, the `tmd_interp` function can provide this sort of output.
