### 2024-01-04 chadagreene/Tide-Model-Driver#12 @lnferris
"may attempt to create your own data file yourself using"

Is the world attempt used because you are not sure if the resulting file is correct, or because it is a time-intensive process?
