### 2024-01-04 chadagreene/Tide-Model-Driver#11 @lnferris
Awesome.

Also I think Conlist is an output of tmd_tide_pred and tmd_extract_HC in TMD 2.5 (so those are sort of the clunky TMD 2.5 proxy of conlist?). It is nice that TMD 3.0 spits it out directly. You could say "output of `tmd_tide_pred` and `tmd_extract_HC`" in the conlist "previous name" box if you wanted.
