### 2024-01-04 a3727af
Update tide_model_data.md
Addressing Issue #12.
