### 2024-01-04 chadagreene/Tide-Model-Driver#12 @chadagreene
Ha, I see what you mean. The intention was only to convey that it's time intensive, but I understand the ambiguity introduced by the word "attempt". I've fixed it now.
