### 2024-01-08 chadagreene/Tide-Model-Driver#20 @chadagreene
Ahh, okay, so it's just a matter of radians vs degrees. I see now that I hadn't included the angle units in the documentation.

I've just updated all the documentation to be clear that the angle units are radians. Thanks for catching this.
