### 2024-01-08 chadagreene/Tide-Model-Driver#19 @lnferris
These functions do not meet the bar for `stream3`, and the expectation for documentation of open-source software (including under-the-hood functions) is that which allows a future user to modify or contribute to software. This revision of TMD was presented as addressing prior clarity issues, one such being undocumented magic numbers  and lack of documentation in under-the-hood functions.
