### 2024-01-08 chadagreene/Tide-Model-Driver#19 @lnferris
For example, the outputs of `tmd_harp` says:
```
% Outpus:
%    hhat: reconstructed tide (size Tx1).
```

The function body of `tmd_astrol` includes:

```
t_mjd = t - datenum(1992,1,1) + 48622;

% Convert time from UTC in decimal MJD to ???
T = t_mjd - 51544.4993;

% Mean longitude of moon:
s = 218.3164 + 13.17639648 * T;
```
