### 2024-01-08 chadagreene/Tide-Model-Driver#19 @chadagreene
Note, I was referencing `stream3c`, not `stream3`.

These time conversions, including the corresponding comments, are the exact code that has been part of TMD for decades, and the conversions occur inside a function that users should never need to directly interact with. Our approach in this update to TMD has been to keep the exact calculations that have stood the test of time, but write the user-facing functions in such a way that users never have to think about the conversions.
