### 2024-01-08 #6018 @kthyng
@chadagreene The reviewer should close the issues brought up in this review. I don't have permission to open the issues myself, but @lnferris could you decide if you agree that the following are closed?
* https://github.com/chadagreene/Tide-Model-Driver/issues/19
* https://github.com/chadagreene/Tide-Model-Driver/issues/20
* https://github.com/chadagreene/Tide-Model-Driver/issues/8

In particular I see my opinion is requested by @lnferris regarding https://github.com/chadagreene/Tide-Model-Driver/issues/8 so please reopen and I'll read through the comments and manuscript as soon as I can.

Thank you!
