### 2024-01-10 chadagreene/Tide-Model-Driver#20 @lnferris
Can you mention this in your [table](https://github.com/chadagreene/Tide-Model-Driver/blob/main/doc/whats_new.md) because a subtle units change has the potential to break people's tools? I agree that radians is better but had to dig to find this difference.
