### 2024-01-10 #6018 @lnferris
@kthyng I think Chad can close 20 himself after completing the quick task in that issue.

My opinions regarding #8 and #19 have not changed. The software is great but the paper is devoid of information relative to other JOSS oceanography papers (e.g. oceananigans, oceanspy...) and there are some under-the-hood-but-just-barely-so functions that are not fully documented. If you think it is close enough for horseshoes than that is fine with me, up to you.
