### 2024-01-10 5c57287
Update whats_new.md
Closing Issue #20.
