### 2024-01-11 #6018 @chadagreene
Thanks @lnferris. You're a tough reviewer, but I really do appreciate the scrutiny. I come away with added confidence in the toolbox knowing that it has (mostly) passed your inspection.

### Issue 8
The paper has met all of the criteria in the [JOSS guidelines](https://joss.readthedocs.io/en/latest/review_criteria.html) which state:

> ...the JOSS paper (the compiled PDF associated with this submission) should only include:
>
> * A list of the authors of the software and their affiliations.
> * A summary describing the high-level functionality and purpose of the software for a diverse, non-specialist audience.
> * A clear statement of need that illustrates the purpose of the software.
> * A description of how this software compares to other commonly-used packages in this research area.
> * Mentions (if applicable) of any ongoing research projects using the software or recent scholarly publications enabled by it.
> * A list of key references including a link to the software archive.

We have provided all of the above in the paper. The reviewer has requested that we include examples in the paper, but the JOSS guidelines also state:

> Note the paper should not include software documentation...

Examples and other forms of documentation belong in the living repo, not in the static paper. The paper describes what TMD does, gives a brief overview of its history, and points readers to the repo, where they can explore the intuitive, extensive, and I'd say downright beautiful documentation.

### Issue 19
We've done as much as we can on this one. We've expanded the documentation of these functions from previous (almost entirely undocumented) versions of TMD, even though these functions are not user-facing. Users should never need to directly interact with these functions in any way, but we've documented them as well as we could for posterity.
