### 2024-01-11 chadagreene/Tide-Model-Driver#8 @kthyng
I'll respond here about the JOSS paper contents. I agree with @chadagreene that JOSS papers are meant to be quite concise. @lnferris you're right that often they are not — JOSS editors have been purposefully pretty liberal with our interpretation of what is ok, preferring to let the authors choose, within the scope of the JOSS guidelines.

That said, @chadagreene can you change the links in lines 22 and 23 to be markdown links pointing to those addresses, or otherwise format them in some way to look like URLs?
