### 2024-01-11 chadagreene/Tide-Model-Driver#19 @kthyng
I understand @chadagreene's reasoning for not changing the code since part of the goal here is apparently to keep that the same over time. I don't see why you shouldn't add more documentation to make things clearer, especially when specifically requested. @lnferris's reasoning is good that that is part of the point of having open source code and there is an extra level of expectation when publishing open source code. The answers to the questions above could be notes in the function(s).
