### 2024-01-11 #6018 @kthyng
I commented in https://github.com/chadagreene/Tide-Model-Driver/issues/8 that I agree this issue can be closed regarding the paper content since JOSS papers can be unusually concise, though first I had a request to @chadagreene in that issue.

Also commented in https://github.com/chadagreene/Tide-Model-Driver/issues/19 and I agree there is more work to do there (unless some of this has already happened?).
