### 2024-01-11 45b0f7b
Update paper.md
Resolving Issue #8.
