+++ paper/paper.md
@@ -40,7 +40,7 @@ bibliography: paper.bib
 
 # Summary
 
-Astronomically-forced tides influence ocean surface height and currents on timescales of minutes to years. Tides contribute to ocean mixing [@munk:1998] and mean flows [@loder:1980], ice sheet and sea ice dynamics, and melting of ice shelves and marine terminating glaciers [@padman:2018]. Tidal signals must be accurately removed when calculating long-term trends in ocean surface height from tidally aliased satellite altimetry measurements [@smith:2000], and the tidal component of ocean circulation must be removed for analyses of ship-based current measurements [@carrillo:2005]. Several models have been made publicly available to predict tides on global [@stammer:2014] or regional [e.g., @padman:2004] scales. Each model contains only information about the complex coefficients of tidal constituents, and therefore requires software to extract and manipulate the data into values of tidal height or transport. The MATLAB package presented here is available at https://github.com/chadagreene/Tide-Model-Driver and offers an alternative to similar programs that are available in Fortran and Python (available at https://www.tpxo.net/otps and https://github.com/tsutterley/pyTMD, respectively). 
+Astronomically-forced tides influence ocean surface height and currents on timescales of minutes to years. Tides contribute to ocean mixing [@munk:1998] and mean flows [@loder:1980], ice sheet and sea ice dynamics, and melting of ice shelves and marine terminating glaciers [@padman:2018]. Tidal signals must be accurately removed when calculating long-term trends in ocean surface height from tidally aliased satellite altimetry measurements [@smith:2000], and the tidal component of ocean circulation must be removed for analyses of ship-based current measurements [@carrillo:2005]. Several models have been made publicly available to predict tides on global [@stammer:2014] or regional [e.g., @padman:2004] scales. Each model contains only information about the complex coefficients of tidal constituents, and therefore requires software to extract and manipulate the data into values of tidal height or transport. The MATLAB package presented here is available at [https://github.com/chadagreene/Tide-Model-Driver](https://github.com/chadagreene/Tide-Model-Driver) and offers an alternative to similar programs that are available in Fortran and Python (available at [https://www.tpxo.net/otps](https://www.tpxo.net/otps) and [https://github.com/tsutterley/pyTMD](https://github.com/tsutterley/pyTMD), respectively). 
 
 # Statement of need
 The Tide Model Driver for MATLAB version 3.0 (TMD3.0) allows users to access and calculate tidal predictions from model coefficients at arbitrary locations and times, and this version of the software represents decades of development by the tide community. The underlying equations for TMD were originally written in Fortran by Richard Ray, and were converted into MATLAB functions by Oregon State University and Earth and Space Research in 2005 [@padman:2022]. The MATLAB version of TMD developed a global user base well before the advent of GitHub or modern documentation standards, but no major updates have been implemented since its inception. The updated toolbox presented here has been restructured for computational efficiency and ease of use, but relies on the same mathematical equations [e.g., @foreman:1989] that have been employed since its earliest implementation. The documentation has been greatly expanded to include clear descriptions of syntax along with many thoroughly explained, replicable examples that use real-world ocean data provided with the software. For TMD3.0, we also introduce a new, consolidated NetCDF tide model data format that is compact, user friendly, and can be used for any barotropic ocean or load tide model. TMD3.0 is issued as a free, open-source software package that is available to all. 
