### 2024-02-06 e51f393
Update tmd_nodal.m
Updates in response to Issue #19.
