### 2024-02-06 chadagreene/Tide-Model-Driver#19 @chadagreene
> [pu,pf] = tmd_nodal(constituents,p,N): what are pu, pf?

We have clarified "Nodal correction outputs pu,pf have units of radians and are are MxN dimensions, corresponding to M timesteps and N constituents."

> tmd_ll2ps(lat,lon,SLAT,SLON,HEMI): allowed inputs (vector, array, etc.)?

Any input dimensions are allowed. By convention, we only specify dimension limitations when they are in fact limited, or when some meaningful insight can be gained through describing the dimensions. Neither is the case here, so we keep it unchanged.

> tmd_harp(t,hc,constituents,p,N,ph,omega): physical units of input/outputs?

Output hhat is the reconstructed tide, as described in the function header, and has the same units accordingly. Input t is described in the function header as having datum units, hc contains the complex harmonic constants and will of course have whatever units the complex harmonic constants have, p is the lunar perigee, N is the lunar node, and phase is in radians as is consistent throughout the TMD3.0 package, and the units of omega are specified as s^-1 in the function header.

> dh = tmd_InferMinor(hc,constituents,t,s,h,p,N): what is a "correction"?

This is a question about tide theory, and if the concept is not understood, then this does not feel like the appropriate place to provide such a lesson. The tmd_InterMinor function is not user-facing, and no one should ever need to open it or even call it directly. The only time anyone should ever need to open this function is if they wish to adapt it into something else, and we assume that anyone who is ever tasked with altering code for minor constituent corrections will understand what minor constituent corrections are. We have described what the function does in the header, we point users to Richard Ray's perth2 function that it is based on, and we described all of the changes we made to the code in this update. No further changes are warranted here.

> tmd_astrol.m: why are there question marks on line 51? Given there are several time formats within TMD stack code: days since Jan 1, 1992; days since Nov 17, 1858, and MATLAB datenum (days since 00-Jan-0000'); the "???" does not inspire user confidence.

The question marks are preserved intact from the legacy code that we have adapted, and we haven't touched the code or time formats, so it doesn't seem appropriate to erase the comments from previous coders.
