### 2024-02-19 #6018 @lnferris
The author did not really make the documentation additions but it is your call at this point. The software is good. Thanks!

> On Feb 19, 2024, at 10:04, Kristen Thyng ***@***.***> wrote:
>
> ﻿
> Hi @lnferris Sorry to bug but I think we're almost to the finish line for this one. Just need to hear back from you.
>
