### 2024-02-19 #6018 @kthyng
@chadagreene I didn't realize that you wrote stuff out in this [issue](https://github.com/chadagreene/Tide-Model-Driver/issues/19) but didn’t change much in the code itself — I don't usually have to look so closely. The fact that the reviewer has asked for clarification on your software documentation indicates that it is not as clear as it could be. I will go make a detailed response to you in the issue.

Thank you for time looking @lnferris.
