### 2024-02-20 72caf63
documentation
Function header documentation edits to address Issue #19.
