### 2024-02-20 chadagreene/Tide-Model-Driver#19 @chadagreene
> You have a typo in this section.

Fixed.

> Explicit is better than implicit and this has the potential to create confusion. Please update.

Done.

> Explicit is better than implicit. If you can write it out here to improve clarity, you can add it to the docs to improve clarity.

Done.

> You can still say something useful. For example: you can add to the docs that the function is not user facing but also be aware that users look at the code and add more detail to improve your codebase. Perhaps you can briefly summarize the correction or you can link to something that provides a longer explanation.

Perhaps it's best to let the original author of the code describe it in his own words. In the function header I've added the note that "tmd_InferMinor returns correction for 16 minor tidal constituents... This is based on Richard Ray's code perth2. For further description, see Ray, Richard D. "Global ocean tide models on the eve of TOPEX/POSEIDON." IEEE transactions on Geoscience and Remote Sensing 31.2 (1993): 355-364."

> This is your chance to improve the comments in the code. If you can figure out what the "???" should be, you should fix the comment.

I've removed the question marks.
