### 2024-02-20 5712036
Update whats_new.md
Fixing a couple of hyperlinks.
