+++ doc/whats_new.md
@@ -3,11 +3,11 @@
 # What's new in TMD3.0?
 TMD3.0 is a rewrite of the [Tide Model Driver for MATLAB v2.5](https://github.com/EarthAndSpaceResearch/TMD_Matlab_Toolbox_v2.5). The goal of this update was to create a clean, efficient set of functions that are well documented, easy to use, and easy to debug. We also wanted to reduce the size, complexity, and overall hassle of dealing with multi-file or binary tide model data files. The biggest changes in this update to TMD3.0 are as follows: 
 
-* Introduced a new, [consolidated NetCDF model data format](doc/TMD_model_file_format.md).  
+* Introduced a new, [consolidated NetCDF model data format](TMD_model_file_format.md).  
 * All functions rewritten for improved performance (*much faster* and lower memory usage than TMD2.5). 
 * Most functions were renamed and inputs have changed to make them more intuitive and easier to use.
 * All new documentation.
-* Updated the CATS2008 model, as described [here](doc/cats2008_updates.md).
+* Updated the CATS2008 model, as described [here](cats2008_updates.md).
 
 # Converting from previous TMD releases
 Here's a table of how the new functions compare to previous versions of TMD: 
