### 2024-02-24 #6018 @lnferris
I am having Github 2FA issues but the remaining boxes can be checked. I have no further comments.

> On Feb 21, 2024, at 5:32 PM, Kristen Thyng ***@***.***> wrote:
>
>
> Ok @lnferris <https://github.com/lnferris> has 3 remaining checkboxes. I have ended up interceding on two of them and will note here that they have been accounted for with the updates in chadagreene/Tide-Model-Driver#19 <https://github.com/chadagreene/Tide-Model-Driver/issues/19>.
>
> Functionality documentation: Is the core functionality of the software documented to a satisfactory level (e.g., API method documentation)?
>
> Quality of writing: Is the paper well written (i.e., it does not require editing for structure, language, or writing quality)?
>
> @lnferris <https://github.com/lnferris> Did you have any further comments or questions on
>
> Functionality: Have the functional claims of the software been confirmed?
>
