### 2024-02-27 #6018 @kthyng
https://joss.readthedocs.io/en/latest/submitting.html#submission-requirements
