### 2024-03-04 #6018 @kthyng
@chadagreene do you want to keep v3.0.0 associated with this JOSS paper?
