### 2024-03-04 #6018 @kthyng
I see it is v3.0.0 on github but v1 on Zenodo
