### 2024-03-04 #6018 @kthyng
Also can you update the metadata on Zenodo to include all of the authors from your JOSS paper?
