### 2024-03-08 #6018 @chadagreene
@kthyng Thanks for catching these things. I've filled in the author list on Zenodo and made the version 3.0.0 consistent throughout.
